using an instance instead of :new() everytime
